Get correct story_title from article HTML
Get the correct story_title from the body of the scraped article.

Previously, item[story_title] was set by finding the first div
with "dc:title" attribute. This led to incorrect behavior, as story
title would be assigned the value of the text in the first "dc:title"
attribute, usually one of the links in the header menu.

This change now assigns story_title the value of the text in the first
"dc:title" attribute in the ** section#section-content **, which is
the desired behavior of finding the story's title.